chore: Move `service/util.rs` into its own module and separate `run_parallel_tasks`
