+++ crates/service/src/util/mod.rs
@@ -3,20 +3,14 @@
 // Copyright: 2024, Rémi Bardon <remi@remibardon.name>
 // License: Mozilla Public License v2.0 (MPL v2.0)
 
-use std::{cmp::min, future::Future};
-
-use futures::{
-    stream::{FuturesOrdered, FuturesUnordered},
-    StreamExt,
-};
-use tokio::{sync::mpsc, task::JoinHandle, time::sleep};
-use tokio_util::sync::CancellationToken;
-use tracing::{debug, error};
+mod parallel_task_runner;
 
 use crate::{
     models::jid::{self, BareJid, NodePart, JID},
     server_config::ServerConfig,
 };
+
+pub use self::parallel_task_runner::*;
 
 pub fn to_bare_jid(jid: &JID) -> Result<BareJid, jid::Error> {
     BareJid::new(jid.to_string().as_str())
@@ -151,87 +145,4 @@ macro_rules! sea_orm_string_enum {
             Some(Self::iter().map(|v| v.to_string().len()).max().unwrap() as u32)
         );
     };
-}
-
-enum Futures<F: Future> {
-    Ordered(FuturesOrdered<F>),
-    Unordered(FuturesUnordered<F>),
-}
-
-impl<F: Future> Futures<F> {
-    fn new(iter: impl Iterator<Item = F>, ordered: bool) -> Self {
-        if ordered {
-            Self::Ordered(iter.collect())
-        } else {
-            Self::Unordered(iter.collect())
-        }
-    }
-    async fn next(&mut self) -> Option<F::Output> {
-        match self {
-            Futures::Ordered(futures) => futures.next().await,
-            Futures::Unordered(futures) => futures.next().await,
-        }
-    }
-}
-
-impl<F: Future> From<FuturesOrdered<F>> for Futures<F> {
-    fn from(futures: FuturesOrdered<F>) -> Self {
-        Self::Ordered(futures)
-    }
-}
-impl<F: Future> From<FuturesUnordered<F>> for Futures<F> {
-    fn from(futures: FuturesUnordered<F>) -> Self {
-        Self::Unordered(futures)
-    }
-}
-
-pub fn run_parallel_tasks<F, R>(
-    futures: Vec<F>,
-    on_cancel: impl FnOnce() -> () + Send + 'static,
-    timeout: std::time::Duration,
-    ordered: bool,
-) -> mpsc::Receiver<R>
-where
-    F: Future<Output = R> + Send + Unpin + 'static,
-    R: Send + 'static,
-{
-    let (tx, rx) = mpsc::channel::<R>(min(futures.len(), 32));
-    tokio::spawn(async move {
-        let cancellation_token = CancellationToken::new();
-        let mut tasks: Futures<JoinHandle<Option<R>>> = Futures::new(
-            futures.into_iter().map(|future| {
-                let cancellation_token = cancellation_token.clone();
-                tokio::spawn(async move {
-                    tokio::select! {
-                        res = future => { Some(res) },
-                        _ = cancellation_token.cancelled() => { None },
-                    }
-                })
-            }),
-            ordered,
-        );
-
-        tokio::select! {
-            _ = async {
-                // NOTE: If `futures.len() == 0` then this `tokio::select!` ends instantly.
-                while let Some(Ok(Some(msg))) = tasks.next().await {
-                    if let Err(err) = tx.send(msg).await {
-                        if tx.is_closed() {
-                            debug!("Cannot send task result: Task aborted.");
-                        } else {
-                            error!("Cannot send task result: {err}");
-                        }
-                    }
-                }
-            } => {}
-            _ = sleep(timeout) => {
-                debug!("Timed out. Cancelling all tasks…");
-
-                cancellation_token.cancel();
-                on_cancel();
-            }
-        };
-    });
-
-    rx
 }
